added front end code to create and delete projects

+++ public/javascripts/modules/hapyak.js
@@ -5,15 +5,50 @@ window.$ = $;
 $(document).ready(function() {
   $("#hapyCreateProj").click(e => handleCreateProject(e));
   $("#hapyDeleteProj").click(e => handleDeleteProject(e));
+  $("#hapyakProjects").on("click", "tr", function() {
+    $(this).toggleClass("selected");
+  });
 
-  let handleCreateProject = e => {
+  const projectList = $("#hapyakProjects").DataTable({
+    dom: "lfrtBip",
+    buttons: [
+      {
+        extend: "csvHtml5",
+        text: "Download CSV",
+        className: "btn btn-default",
+        filename: "Hapyak_Projects"
+      }
+    ],
+    ajax: "/listProjects",
+    columns: [
+      { data: "_id" },
+      { data: "id" },
+      { data: "title" },
+      { data: "video" },
+      { data: "brightcoveId" },
+      { data: "track" },
+      { data: "created" }
+    ],
+    columnDefs: [
+      {
+        targets: [0],
+        visible: false,
+        searchable: true
+      }
+    ],
+    pageLength: 25
+  });
+
+  function handleCreateProject(e) {
     console.log("Project Created!");
+
+    const videoId = $("#bcVideoId").val();
 
     $.ajax({
       url: `/createProject`,
       type: "POST",
       data: {
-        say: "Hello"
+        video_source_id: videoId
       }
     })
       .done(res => {
@@ -22,23 +57,63 @@ $(document).ready(function() {
       .fail(err => {
         console.log(err);
       });
-  };
 
-  let handleDeleteProject = e => {
+    projectList.ajax.reload();
+  }
+
+  function handleDeleteProject(e) {
     console.log("Project Deleted!");
 
-    $.ajax({
-      url: `/deleteProject`,
-      type: "DELETE",
-      data: {
-        say: "Hello"
+    const projects = projectList.rows(".selected").data();
+    const msg = `Are you sure you want to delete ${projects.length} projects?`;
+
+    swal({
+      title: "Are you sure?",
+      text: msg,
+      type: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "Cancel",
+      confirmButtonClass: "btn btn-success",
+      cancelButtonClass: "btn btn-danger",
+      buttonsStyling: false
+    }).then(result => {
+      if (result.value) {
+        for (let i = 0; i < projects.length; i++) {
+          $.ajax({
+            url: "/deleteProject",
+            dataType: "json",
+            type: "DELETE",
+            contentType: "application/json",
+            data: JSON.stringify({ toBeDeleted: projects[i].id })
+          });
+        }
+
+        swal({
+          position: "top",
+          type: "success",
+          title: "Recording(s) deleted!",
+          showConfirmButton: false,
+          timer: 800,
+          buttonsStyling: false
+        }).then(result => {
+          projectList
+            .rows(".selected")
+            .remove()
+            .draw(false);
+        });
+        // result.dismiss can be 'cancel', 'overlay',
+        // 'close', and 'timer'
+      } else if (result.dismiss) {
+        swal({
+          position: "top",
+          type: "error",
+          title: "Canceled!",
+          showConfirmButton: false,
+          timer: 800,
+          buttonsStyling: false
+        });
       }
-    })
-      .done(res => {
-        console.log(res);
-      })
-      .fail(err => {
-        console.log(err);
-      });
-  };
+    });
+  }
 });
